feat: implement cache (that needs to be refactored, it's ported 1:1 from python code)

+++ src/models/creature.rs
@@ -1,12 +1,12 @@
 use crate::models::enums::{AlignmentEnum, RarityEnum, SizeEnum};
 use serde::{Deserialize, Serialize};
 
-#[derive(Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct Creature {
     pub id: String,
     pub name: String,
     pub hp: i16,
-    pub level: i16,
+    pub level: i8,
     pub alignment: AlignmentEnum,
     pub size: SizeEnum,
     pub family: Option<String>,
